:sparkles: LoS now detects and calls `lose_level` on lose
