+++ game/assets/tilemaps/TestLevel.tmx
@@ -151,15 +151,7 @@
   <object id="38" name="spawn" type="spawn" x="56" y="88" width="8" height="8"/>
   <object id="39" name="1" type="point" x="48" y="88" width="8" height="8"/>
  </objectgroup>
- <objectgroup color="#0000ff" id="9" name="guard6">
-  <properties>
-   <property name="type" value="guard"/>
-  </properties>
-  <object id="40" name="0" type="point" x="112" y="96" width="8" height="8"/>
-  <object id="41" name="spawn" type="spawn" x="72" y="96" width="8" height="8"/>
-  <object id="42" name="1" type="point" x="48" y="96" width="8" height="8"/>
- </objectgroup>
- <objectgroup color="#0000ff" id="10" name="guard7">
+ <objectgroup color="#0000ff" id="10" name="guard6">
   <properties>
    <property name="type" value="guard"/>
   </properties>
